+++ packages/firebase_auth/test/firebase_auth_test.dart
@@ -25,6 +25,25 @@ const String kMockPhoneNumber = '5555555555';
 const String kMockVerificationId = '12345';
 const String kMockSmsCode = '123456';
 const String kMockLanguage = 'en';
+const Map<String, dynamic> kMockAdditionalUserInfo = <String, dynamic>{
+  'isNewUser': false,
+  'username': 'flutterUser',
+  'providerId': 'testProvider',
+  'profile': <String, dynamic>{'foo': 'bar'},
+};
+const Map<String, dynamic> kMockUser = <String, dynamic>{
+  'isAnonymous': true,
+  'isEmailVerified': false,
+  'providerData': <Map<String, String>>[
+    <String, String>{
+      'providerId': kMockProviderId,
+      'uid': kMockUid,
+      'displayName': kMockDisplayName,
+      'photoUrl': kMockPhotoUrl,
+      'email': kMockEmail,
+    },
+  ],
+};
 
 void main() {
   group('$FirebaseAuth', () {
@@ -48,6 +67,8 @@ void main() {
           case "startListeningAuthState":
             return mockHandleId++;
             break;
+          case "currentUser":
+            return kMockUser;
           case "sendLinkToEmail":
           case "sendPasswordResetEmail":
           case "updateEmail":
@@ -63,7 +84,10 @@ void main() {
             return null;
             break;
           default:
-            return mockFirebaseUser();
+            return <String, dynamic>{
+              'user': kMockUser,
+              'additionalUserInfo': kMockAdditionalUserInfo,
+            };
             break;
         }
       });
@@ -81,27 +105,25 @@ void main() {
       expect(userInfo.email, kMockEmail);
     }
 
-    test('currentUser', () async {
-      final FirebaseUser user = await auth.currentUser();
-      verifyUser(user);
+    void verifyAuthResult(AuthResult result) {
+      verifyUser(result.user);
+      final AdditionalUserInfo additionalUserInfo = result.additionalUserInfo;
       expect(
-        log,
-        <Matcher>[
-          isMethodCall('currentUser',
-              arguments: <String, String>{'app': auth.app.name}),
-        ],
-      );
-    });
+          additionalUserInfo.isNewUser, kMockAdditionalUserInfo['isNewUser']);
+      expect(additionalUserInfo.username, kMockAdditionalUserInfo['username']);
+      expect(
+          additionalUserInfo.providerId, kMockAdditionalUserInfo['providerId']);
+      expect(additionalUserInfo.profile, kMockAdditionalUserInfo['profile']);
+    }
 
-    test('signInAnonymously', () async {
-      final FirebaseUser user = await auth.signInAnonymously();
-      verifyUser(user);
+    test('getIdToken', () async {
+      final FirebaseUser user = await auth.currentUser();
       expect(await user.getIdToken(), equals(kMockIdToken));
       expect(await user.getIdToken(refresh: true), equals(kMockIdToken));
       expect(
         log,
         <Matcher>[
-          isMethodCall('signInAnonymously',
+          isMethodCall('currentUser',
               arguments: <String, String>{'app': auth.app.name}),
           isMethodCall(
             'getIdToken',
@@ -114,6 +136,18 @@ void main() {
             'getIdToken',
             arguments: <String, dynamic>{'refresh': true, 'app': auth.app.name},
           ),
+        ],
+      );
+    });
+
+    test('signInAnonymously', () async {
+      final AuthResult result = await auth.signInAnonymously();
+      verifyAuthResult(result);
+      expect(
+        log,
+        <Matcher>[
+          isMethodCall('signInAnonymously',
+              arguments: <String, String>{'app': auth.app.name}),
         ],
       );
     });
@@ -158,10 +192,11 @@ void main() {
     });
 
     test('signInWithEmailAndLink', () async {
-      await auth.signInWithEmailAndLink(
+      final AuthResult result = await auth.signInWithEmailAndLink(
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -175,11 +210,11 @@ void main() {
     });
 
     test('createUserWithEmailAndPassword', () async {
-      final FirebaseUser user = await auth.createUserWithEmailAndPassword(
+      final AuthResult result = await auth.createUserWithEmailAndPassword(
         email: kMockEmail,
         password: kMockPassword,
       );
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -219,9 +254,9 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -251,8 +286,8 @@ void main() {
         email: 'test@example.com',
         link: '<Url with domain from your Firebase project>',
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -302,9 +337,9 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -334,8 +369,8 @@ void main() {
         authToken: kMockIdToken,
         authTokenSecret: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -358,9 +393,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -388,8 +423,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -412,9 +447,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -444,8 +479,8 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -469,7 +504,8 @@ void main() {
         verificationId: kMockVerificationId,
         smsCode: kMockSmsCode,
       );
-      await auth.signInWithCredential(credential);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(log, <Matcher>[
         isMethodCall('signInWithCredential', arguments: <String, dynamic>{
           'app': auth.app.name,
@@ -508,7 +544,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -533,7 +571,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -558,7 +598,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -583,7 +625,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -608,7 +652,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      await user.reauthenticateWithCredential(credential);
+      final AuthResult result =
+          await user.reauthenticateWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -631,9 +677,9 @@ void main() {
         idToken: kMockIdToken,
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -662,9 +708,9 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -692,8 +738,8 @@ void main() {
       final AuthCredential credential = FacebookAuthProvider.getCredential(
         accessToken: kMockAccessToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -716,9 +762,9 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -748,8 +794,8 @@ void main() {
         authToken: kMockAuthToken,
         authTokenSecret: kMockAuthTokenSecret,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -772,9 +818,9 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -802,8 +848,8 @@ void main() {
       final AuthCredential credential = GithubAuthProvider.getCredential(
         token: kMockGithubToken,
       );
-      final FirebaseUser user = await auth.signInWithCredential(credential);
-      verifyUser(user);
+      final AuthResult result = await auth.signInWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -826,9 +872,9 @@ void main() {
         email: kMockEmail,
         password: kMockPassword,
       );
-      FirebaseUser user = await auth.currentUser();
-      user = await user.linkWithCredential(credential);
-      verifyUser(user);
+      final FirebaseUser user = await auth.currentUser();
+      final AuthResult result = await user.linkWithCredential(credential);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1122,9 +1168,9 @@ void main() {
     });
 
     test('signInWithCustomToken', () async {
-      final FirebaseUser user =
+      final AuthResult result =
           await auth.signInWithCustomToken(token: kMockCustomToken);
-      verifyUser(user);
+      verifyAuthResult(result);
       expect(
         log,
         <Matcher>[
@@ -1163,7 +1209,7 @@ void main() {
       await Future<void>.delayed(const Duration(seconds: 0));
 
       await simulateEvent(null);
-      await simulateEvent(mockFirebaseUser());
+      await simulateEvent(kMockUser);
 
       final FirebaseUser user1 = await events.remove();
       expect(user1, isNull);
@@ -1210,26 +1256,6 @@ void main() {
     });
   });
 }
-
-Map<String, dynamic> mockFirebaseUser(
-        {String providerId = kMockProviderId,
-        String uid = kMockUid,
-        String displayName = kMockDisplayName,
-        String photoUrl = kMockPhotoUrl,
-        String email = kMockEmail}) =>
-    <String, dynamic>{
-      'isAnonymous': true,
-      'isEmailVerified': false,
-      'providerData': <Map<String, String>>[
-        <String, String>{
-          'providerId': providerId,
-          'uid': uid,
-          'displayName': displayName,
-          'photoUrl': photoUrl,
-          'email': email,
-        },
-      ],
-    };
 
 /// Queue whose remove operation is asynchronous, awaiting a corresponding add.
 class AsyncQueue<T> {